PixelGrid Project
In this project, through a java compiler, I managed to  create an interactive program which reads in a PNG image into my code, and created copies of said image with alterations. One alteration I did through code was to find the column in the image with the highest blue values in each pixel. I then highlighted the column in pure blue, and deleted it. I also created a random column finder function in my code, which highlighted a random column of pixels in the image and deleted it. The user would have liberty of choice with the interactive menu, and do as many alterations as they would like to the image.